backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic-settings==2.6.1
feedparser==6.0.11
httpx>=0.27.0,<0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
ollama==0.4.4
apscheduler==3.10.4
python-dotenv==1.0.1
groq==0.13.0
psycopg2-binary==2.9.10
google-generativeai==0.8.3
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
